Fixes Applied

+++ lib/repositories/collection_route_repository.dart
@@ -1,10 +1,10 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:zero_waste/models/collection_route.dart';
-import 'package:zero_waste/models/waste_bin.dart';
+import 'package:zero_waste/utils/app_logger.dart';
 
 class CollectionRouteRepository {
   final CollectionReference _routeCollection =
-  FirebaseFirestore.instance.collection('route');
+      FirebaseFirestore.instance.collection('route');
 
   Future<String> generateRouteId() async {
     try {
@@ -18,6 +18,7 @@ class CollectionRouteRepository {
         return userId;
       }
     } catch (e) {
+      AppLogger.printError(e.toString());
       throw Exception('Error generating unique Route ID: $e');
     }
   }
@@ -25,24 +26,44 @@ class CollectionRouteRepository {
   Future<void> addRoute(CollectionRoute collectionRoute) async {
     try {
       collectionRoute.id = await generateRouteId();
-      await _routeCollection.doc(collectionRoute.id).set(collectionRoute.toMap());
+      await _routeCollection
+          .doc(collectionRoute.id)
+          .set(collectionRoute.toMap());
+      AppLogger.printInfo('Route Added successfully.');
     } catch (e) {
+      AppLogger.printError(e.toString());
       throw Exception('Error adding CollectionRoute: $e');
     }
   }
 
   Future<void> updateRoute(CollectionRoute collectionRoute) async {
     try {
-      await _routeCollection.doc(collectionRoute.id).update(collectionRoute.toMap());
+      final doc = await _routeCollection.doc(collectionRoute.id).get();
+      if (!doc.exists) {
+        throw Exception('Route does not exist: ${collectionRoute.id}');
+      } else {
+        await _routeCollection
+            .doc(collectionRoute.id)
+            .update(collectionRoute.toMap());
+        AppLogger.printInfo('Route Updated successfully.');
+      }
     } catch (e) {
-      throw Exception('Error updating CollectionRoute: $e');
+      AppLogger.printError(e.toString());
+      throw Exception('Error updating Collection Route: $e');
     }
   }
 
   Future<void> deleteRoute(String routeId) async {
     try {
-      await _routeCollection.doc(routeId).delete();
+      final doc = await _routeCollection.doc(routeId).get();
+      if (!doc.exists) {
+        throw Exception('Route does not exist: $routeId');
+      } else {
+        await _routeCollection.doc(routeId).delete();
+        AppLogger.printInfo('Route Deleted successfully.');
+      }
     } catch (e) {
+      AppLogger.printError(e.toString());
       throw Exception('Error deleting CollectionRoute: $e');
     }
   }
@@ -50,8 +71,11 @@ class CollectionRouteRepository {
   Future<List<CollectionRoute>> getAllRoutes() async {
     try {
       QuerySnapshot snapshot = await _routeCollection.get();
-      return snapshot.docs.map((doc) => CollectionRoute.fromDocument(doc)).toList();
+      return snapshot.docs
+          .map((doc) => CollectionRoute.fromDocument(doc))
+          .toList();
     } catch (e) {
+      AppLogger.printError(e.toString());
       throw Exception('Error getting all CollectionRoute: $e');
     }
   }
@@ -62,9 +86,11 @@ class CollectionRouteRepository {
       if (snapshot.exists) {
         return CollectionRoute.fromDocument(snapshot);
       } else {
-        throw Exception('CollectionRoute not found');
+        AppLogger.printError('Collection Route not found');
+        throw Exception('Collection Route not found');
       }
     } catch (e) {
+      AppLogger.printError(e.toString());
       throw Exception('Error getting CollectionRoute: $e');
     }
   }
