+++ lib/repositories/household_user_repository.dart
@@ -2,7 +2,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:zero_waste/models/household_user.dart';
 import 'package:zero_waste/models/user.dart';
 import 'package:zero_waste/repositories/user_repository.dart';
-import 'package:firebase_storage/firebase_storage.dart';
+import 'package:zero_waste/utils/app_logger.dart';
 
 class HouseholdUserRepository {
   final CollectionReference _usersCollection =
@@ -20,6 +20,7 @@ class HouseholdUserRepository {
         return userId;
       }
     } catch (e) {
+      AppLogger.printError(e.toString());
       throw Exception('Error generating unique user ID: $e');
     }
   }
@@ -30,23 +31,41 @@ class HouseholdUserRepository {
       householdUser.userId = userId.id;
       householdUser.id = await generateUniqueId();
       await _usersCollection.doc(householdUser.id).set(householdUser.toMap());
+      AppLogger.printInfo('User Added successfully.');
     } catch (e) {
+      AppLogger.printError(e.toString());
       throw Exception('Error adding HouseholdUser: $e');
     }
   }
 
   Future<void> updateUser(String userId, HouseholdUser user) async {
     try {
-      await _usersCollection.doc(userId).update(user.toMap());
+      final doc = await _usersCollection.doc(userId).get();
+      if (!doc.exists) {
+        AppLogger.printError('User does not exist: $userId');
+        throw Exception('User does not exist: $userId');
+      } else {
+        await _usersCollection.doc(userId).update(user.toMap());
+        AppLogger.printInfo('User Updated successfully.');
+      }
     } catch (e) {
+      AppLogger.printError(e.toString());
       throw Exception('Error updating HouseholdUser: $e');
     }
   }
 
   Future<void> deleteUser(String userId) async {
     try {
-      await _usersCollection.doc(userId).delete();
+      final doc = await _usersCollection.doc(userId).get();
+      if (!doc.exists) {
+        AppLogger.printError('User does not exist: $userId');
+        throw Exception('User does not exist: $userId');
+      } else {
+        await _usersCollection.doc(userId).delete();
+        AppLogger.printInfo('User Updated successfully.');
+      }
     } catch (e) {
+      AppLogger.printError(e.toString());
       throw Exception('Error deleting HouseholdUser: $e');
     }
   }
@@ -58,6 +77,7 @@ class HouseholdUserRepository {
           .map((doc) => HouseholdUser.fromDocument(doc))
           .toList();
     } catch (e) {
+      AppLogger.printError(e.toString());
       throw Exception('Error getting all HouseholdUser: $e');
     }
   }
@@ -68,9 +88,11 @@ class HouseholdUserRepository {
       if (snapshot.exists) {
         return HouseholdUser.fromDocument(snapshot);
       } else {
+        AppLogger.printError('HouseholdUser not found');
         throw Exception('HouseholdUser not found');
       }
     } catch (e) {
+      AppLogger.printError(e.toString());
       throw Exception('Error getting HouseholdUser: $e');
     }
   }
@@ -87,8 +109,8 @@ class HouseholdUserRepository {
         return null;
       }
     } catch (e) {
-      print('Error getting HouseholdUser by email: $e');
-      return null;
+      AppLogger.printError(e.toString());
+      throw Exception('Error getting HouseholdUser: $e');
     }
   }
 }
