+++ lib/repositories/user_repository.dart
@@ -1,31 +1,47 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:zero_waste/utils/app_logger.dart';
 import '../models/user.dart';
 
 class UserRepository {
   final CollectionReference _usersCollection =
-  FirebaseFirestore.instance.collection('users');
+      FirebaseFirestore.instance.collection('users');
 
   Future<DocumentReference<Object?>> addUser(User user) async {
     try {
       DocumentReference userRef = await _usersCollection.add(user.toMap());
       return userRef;
     } catch (e) {
+      AppLogger.printError(e.toString());
       throw Exception('Error adding user: $e');
     }
   }
 
   Future<void> updateUser(String userId, User user) async {
     try {
-      await _usersCollection.doc(userId).update(user.toMap());
+      final doc = await _usersCollection.doc(userId).get();
+      if (!doc.exists) {
+        throw Exception('User does not exist: $userId');
+      } else {
+        await _usersCollection.doc(userId).update(user.toMap());
+        AppLogger.printInfo('User Updated successfully.');
+      }
     } catch (e) {
+      AppLogger.printError(e.toString());
       throw Exception('Error updating user: $e');
     }
   }
 
   Future<void> deleteUser(String userId) async {
     try {
-      await _usersCollection.doc(userId).delete();
+      final doc = await _usersCollection.doc(userId).get();
+      if (!doc.exists) {
+        throw Exception('User does not exist: $userId');
+      } else {
+        await _usersCollection.doc(userId).delete();
+        AppLogger.printInfo('User Deleted successfully.');
+      }
     } catch (e) {
+      AppLogger.printError(e.toString());
       throw Exception('Error deleting user: $e');
     }
   }
@@ -35,6 +51,7 @@ class UserRepository {
       QuerySnapshot snapshot = await _usersCollection.get();
       return snapshot.docs.map((doc) => User.fromDocument(doc)).toList();
     } catch (e) {
+      AppLogger.printError(e.toString());
       throw Exception('Error getting all users: $e');
     }
   }
@@ -48,13 +65,15 @@ class UserRepository {
         throw Exception('User not found');
       }
     } catch (e) {
+      AppLogger.printError(e.toString());
       throw Exception('Error getting user: $e');
     }
   }
 
   Future<User> login(User user) async {
     try {
-      QuerySnapshot querySnapshot = await _usersCollection.where('email', isEqualTo: user.email).get();
+      QuerySnapshot querySnapshot =
+          await _usersCollection.where('email', isEqualTo: user.email).get();
 
       if (querySnapshot.docs.isNotEmpty) {
         DocumentSnapshot documentSnapshot = querySnapshot.docs.first;
@@ -62,15 +81,16 @@ class UserRepository {
         if (documentSnapshot.get('password') == user.password) {
           return User.fromDocument(documentSnapshot);
         } else {
+          AppLogger.printError('Invalid password');
           throw Exception('Invalid password');
         }
       } else {
+        AppLogger.printError('User not found');
         throw Exception('User not found');
       }
     } catch (e) {
+      AppLogger.printError(e.toString());
       throw Exception('Error getting user: $e');
     }
   }
-
-
 }
