+++ lib/repositories/waste_bin_repository.dart
@@ -1,5 +1,6 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:zero_waste/models/waste_bin.dart';
+import 'package:zero_waste/utils/app_logger.dart';
 
 class WasteBinRepository {
   final CollectionReference _binCollection =
@@ -9,22 +10,37 @@ class WasteBinRepository {
     try {
       await _binCollection.add(wasteBin.toMap());
     } catch (e) {
+      AppLogger.printError(e.toString());
       throw Exception('Error adding WasteBin: $e');
     }
   }
 
   Future<void> updateBin(WasteBin wasteBin) async {
     try {
-      await _binCollection.doc(wasteBin.id).update(wasteBin.toMap());
+      final doc = await _binCollection.doc(wasteBin.id).get();
+      if (!doc.exists) {
+        throw Exception('wasteBin does not exist: ${wasteBin.id}');
+      } else {
+        await _binCollection.doc(wasteBin.id).update(wasteBin.toMap());
+        AppLogger.printInfo('wasteBin Updated successfully.');
+      }
     } catch (e) {
+      AppLogger.printError(e.toString());
       throw Exception('Error updating WasteBin: $e');
     }
   }
 
   Future<void> deleteBin(String binId) async {
     try {
-      await _binCollection.doc(binId).delete();
+      final doc = await _binCollection.doc(binId).get();
+      if (!doc.exists) {
+        throw Exception('wasteBin does not exist: $binId');
+      } else {
+        await _binCollection.doc(binId).delete();
+        AppLogger.printInfo('wasteBin Updated successfully.');
+      }
     } catch (e) {
+      AppLogger.printError(e.toString());
       throw Exception('Error deleting WasteBin: $e');
     }
   }
@@ -34,6 +50,7 @@ class WasteBinRepository {
       QuerySnapshot snapshot = await _binCollection.get();
       return snapshot.docs.map((doc) => WasteBin.fromDocument(doc)).toList();
     } catch (e) {
+      AppLogger.printError(e.toString());
       throw Exception('Error getting all WasteBins: $e');
     }
   }
@@ -44,9 +61,11 @@ class WasteBinRepository {
       if (snapshot.exists) {
         return WasteBin.fromDocument(snapshot);
       } else {
+        AppLogger.printError('WasteBin not found');
         throw Exception('WasteBin not found');
       }
     } catch (e) {
+      AppLogger.printError(e.toString());
       throw Exception('Error getting WasteBin: $e');
     }
   }
@@ -58,6 +77,7 @@ class WasteBinRepository {
 
       return schedules.docs.map((doc) => WasteBin.fromDocument(doc)).toList();
     } catch (e) {
+      AppLogger.printError(e.toString());
       throw Exception('Error getting WasteBin: $e');
     }
   }
